requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
PyMySQL==1.1.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
alembic==1.13.1
PyYAML==6.0.1